+++ src/components/task.js
@@ -1,5 +1,5 @@
 import {monthNames} from '../const.js';
-import {formatTime} from '../utils.js';
+import {formatTime} from '../utils/common.js';
 import AbstractComponent from './abstract-class.js';
 
 const createHashtagsMarkup = (hashtags) => {
