+++ src/utils/common.js
@@ -1,8 +1,3 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 const getRandomIntegerFromGap = (min, max) => {
   return Math.floor(min + Math.random() * (max + 1 - min));
 };
@@ -27,30 +22,9 @@ const formatTime = (date) => {
   return `${hours}:${minutes} ${interval}`;
 };
 
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-const createElement = (template) => {
-  const element = document.createElement(`div`);
-  element.innerHTML = template;
-  return element.firstChild;
-};
-
-
 export {
   getRandomIntegerFromGap,
   getRandomArrayItem,
   formatTime,
   castTimeFormat,
-  createElement,
-  RenderPosition,
-  render
 };
